refactor(active-memory): split recall helpers
